+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Yii Framework 2 redis extension Change Log
 
 - Bug #247: `Cache::getValue()` now returns `false` in case of missing key (rhertogh)
 - Enh #249 Added support to set the redis scheme to tls. Add to configuration: `'scheme' => 'tls'` (tychovbh)
+- Enh #253 Added support to set a scheme for keys of Session. Added `Session::$objectType` and `Session::usePrefixInSchema` (githubjeka)
 
 
 2.0.17 January 11, 2022